requirements.txt
torch
pandas
scikit-learn
fastapi
uvicorn
matplotlib
networkx
kafka-python
transformers
openai
numpy
